Extract origin comparison logic to a separate class
